Bug 63185 - Add option to implicitly trust SSL/TLS connections/Disable hostname verification
Factor use of Trust All Socket Factory
Bugzilla Id: 63185

git-svn-id: https://svn.apache.org/repos/asf/jmeter/trunk@1854618 13f79535-47bb-0310-9956-ffa450edef68
